requirements.txt
openai
python-dotenv
pdf2image
opencv-python-headless
pytesseract
PyMuPDF
